Nicer testing for user agent
